Add buffering and batches to AsyncWriter (#2611)
Update AsyncWriter to buffer log entries and batch API calls.

Back pressure is applied by limiting the number of queued API calls.
Errors will now be raised when there are not enough resources.
Errors are indicated using the AsyncWriter#on_error callback.

Pending log entries are sent before the process closes using at_exit.

The previous implementation was making API requests as fast as it could.
This change allows the writer to buffer the entries longer and send in
more batches. Also add configuration for the max number of entries to
buffer, and well as the max byte size to buffer before the batch is sent.

Deprecate suspend and resume, which are less useful now.